Seperated MPIManager
